Add basic state machine and transitions
This is a long way from functional but has a basis for some
theory as to how to handle transitions and states.

Jogging is now jog forward and jog backward as separate states.

I decided to put in individual transition functions.  Right now
they are purely boiler-plate but it seemed like potentially the
transition between states A and B could require something more than
exiting state A and entering state B.  If this is not needed, there
is merely additional verbosity.  The number of transitions is low.

I have not implemented either the eventing mechanisms nor the data
nor the Kalman filter on position nor a variety of mathematical
utility functions to set units.

This remains a WIP.