Robustness improvements
Threading:
So look.  This isn't great.  There appear to be some threading
issues.  They manifest when the gpiozero button debouncer is non-
None.  That's not entirely unsuprising as threads are used to
implement the debouncing and threads do wacky race-conditiony things.

The manifestation was that sometimes buttons or sensor wouldn't
register as having been released which is a pretty big problem.

So I introduce a mutex around all the state checks in the hopes that
this would cause the problem to go away.  It did not.

So I disabled debouncing in the config and then all was well.

I'm going to leave the recursive mutex in since there are some other
times when gpiozero uses threads and as far as I can tell it does
not cause problems.  I've done several thousand button presses and
not yet encountered an unexpected state violation.

Logging:
I cleaned up some of the logs and combined them into some nicer
functions.

I also explicitly track position as unknown when it is unknown.