Implement manual debounce
This commit is a bit messy and complected with the next one.

We were unable to get the debouncing working well so this implements
my own custom debouncing in software in a separate thread.  It also
updates some of the other changes needed to get the program working
in the real control box including a test script for the same.