Simplifications for rotate sensor removed
We couldn't get he rotation sensor to work and had to remove it from
the system.  That means no speed or position detection which requires
the system to work purely on time requiring a fair amount of
coordination with the motor configuration.

It's not ideal.  It will probably work ok.